+++ pcs_test/tier0/lib/cib/test_constraint_ticket.py
@@ -38,11 +38,11 @@ class PrepareOptionsPlainTest(TestCase):
                 "loss-policy": "fence",
                 "rsc": "resourceA",
                 "rsc-role": const.PCMK_ROLE_PROMOTED_PRIMARY,
-                "ticket": "ticket_key",
+                "ticket": "ticket-key",
             },
             self.prepare(
                 {"loss-policy": "fence", "rsc-role": role},
-                "ticket_key",
+                "ticket-key",
                 "resourceA",
             ),
         )
@@ -65,11 +65,11 @@ class PrepareOptionsPlainTest(TestCase):
                 "id": "generated_id",
                 "loss-policy": "fence",
                 "rsc": "resourceA",
-                "ticket": "ticket_key",
+                "ticket": "ticket-key",
             },
             self.prepare(
                 {"loss-policy": "fence", "rsc-role": ""},
-                "ticket_key",
+                "ticket-key",
                 "resourceA",
             ),
         )
@@ -79,7 +79,7 @@ class PrepareOptionsPlainTest(TestCase):
         assert_raise_library_error(
             lambda: self.prepare(
                 {"unknown": "nonsense", "rsc-role": "master"},
-                "ticket_key",
+                "ticket-key",
                 "resourceA",
             ),
             (
@@ -103,7 +103,7 @@ class PrepareOptionsPlainTest(TestCase):
     def test_refuse_bad_role(self, _):
         assert_raise_library_error(
             lambda: self.prepare(
-                {"id": "id", "rsc-role": "bad_role"}, "ticket_key", "resourceA"
+                {"id": "id", "rsc-role": "bad_role"}, "ticket-key", "resourceA"
             ),
         )
         self.report_processor.assert_reports(
@@ -142,11 +142,30 @@ class PrepareOptionsPlainTest(TestCase):
             ]
         )
 
+    def test_refuse_bad_ticket(self, _):
+        assert_raise_library_error(
+            lambda: self.prepare(
+                {
+                    "id": "id",
+                },
+                "bad_ticket",
+                "resourceA",
+            ),
+        )
+        self.report_processor.assert_reports(
+            [
+                fixture.error(
+                    report_codes.BOOTH_TICKET_NAME_INVALID,
+                    ticket_name="bad_ticket",
+                ),
+            ]
+        )
+
     def test_refuse_missing_resource_id(self, _):
         assert_raise_library_error(
             lambda: self.prepare(
                 {"id": "id", "rsc-role": const.PCMK_ROLE_PROMOTED},
-                "ticket_key",
+                "ticket-key",
                 "",
             ),
         )
@@ -165,7 +184,7 @@ class PrepareOptionsPlainTest(TestCase):
         assert_raise_library_error(
             lambda: self.prepare(
                 {"loss-policy": "unknown", "ticket": "T", "id": "id"},
-                "ticket_key",
+                "ticket-key",
                 "resourceA",
             ),
         )
@@ -190,7 +209,7 @@ class PrepareOptionsPlainTest(TestCase):
             "ticket": "T",
             "rsc-role": const.PCMK_ROLE_PROMOTED,
         }
-        ticket_key = "ticket_key"
+        ticket_key = "ticket-key"
         resource_id = "resourceA"
         expected_options = options.copy()
         expected_options.update(
@@ -306,6 +325,16 @@ class PrepareOptionsWithSetTest(TestCase):
             ),
         )
 
+    def test_refuse_bad_ticket(self, _):
+        assert_raise_library_error(
+            lambda: self.prepare({"id": "id", "ticket": "bad_ticket"}),
+            (
+                severities.ERROR,
+                report_codes.BOOTH_TICKET_NAME_INVALID,
+                {"ticket_name": "bad_ticket"},
+            ),
+        )
+
 
 class Element:
     # pylint: disable=too-few-public-methods
@@ -321,14 +350,14 @@ class AreDuplicatePlain(TestCase):
     def setUp(self):
         self.first = Element(
             {
-                "ticket": "ticket_key",
+                "ticket": "ticket-key",
                 "rsc": "resourceA",
                 "rsc-role": const.PCMK_ROLE_PROMOTED_LEGACY,
             }
         )
         self.second = Element(
             {
-                "ticket": "ticket_key",
+                "ticket": "ticket-key",
                 "rsc": "resourceA",
                 "rsc-role": const.PCMK_ROLE_PROMOTED_LEGACY,
             }
@@ -393,8 +422,8 @@ class AreDuplicateWithResourceSet(TestCase):
         mock_have_duplicate_resource_sets.return_value = True
         self.assertTrue(
             ticket.are_duplicate_with_resource_set(
-                Element({"ticket": "ticket_key"}),
-                Element({"ticket": "ticket_key"}),
+                Element({"ticket": "ticket-key"}),
+                Element({"ticket": "ticket-key"}),
             )
         )
 
@@ -404,7 +433,7 @@ class AreDuplicateWithResourceSet(TestCase):
         mock_have_duplicate_resource_sets.return_value = True
         self.assertFalse(
             ticket.are_duplicate_with_resource_set(
-                Element({"ticket": "ticket_key"}),
+                Element({"ticket": "ticket-key"}),
                 Element({"ticket": "X"}),
             )
         )